lime under tab7. not able to create visualization
LIME is running, but I cannot get the tables to appear in streamlit. i think there is some issue with the conversion between arrays and df columns. I have tried converting all attributes to dataframes to be able to use st.table, but was not successful. if there is an easy way to get this analysis visible in our app, that'd be great, otherwise i think we will need to rely on accuracy and loss to determine the validity of our model, and that won't provide the same insight